+++ server.js
@@ -5,6 +5,7 @@ const routes = require('./routes');
 const PORT = process.env.PORT || 3001;
 const mongoose = require('mongoose');
 const helmet = require('helmet');
+//const stripe = require('stripe')(process.env.SECRET_KEY);
 
 server.use(helmet());
 server.use(
@@ -17,6 +18,26 @@ server.use(express.json());
 if (process.env.NODE_ENV === 'production') {
   server.use(express.static('client/build'));
 }
+
+// const calculateOrderAmount = () => {
+//   // Replace this constant with a calculation of the order's amount
+//   // Calculate the order total on the server to prevent
+//   // people from directly manipulating the amount on the client
+//   return 1400;
+// };
+
+
+// server.post("/payment", async (req, res) => {
+//   const { items } = req.body;
+//   // Create a PaymentIntent with the order amount and currency
+//   const paymentIntent = await stripe.paymentIntents.create({
+//     amount: calculateOrderAmount(items),
+//     currency: "usd"
+//   });
+// res.send({
+//     clientSecret: paymentIntent.client_secret
+//   });
+// });
 
 // Add routes, both API and view
 server.use(routes);
